Updating naming of class and objects
