+++ apps/Vision/HeatMatrix.py
@@ -1,7 +1,7 @@
 import numpy as np
 from queue import Queue
 
-class DirectionVector:
+class HeatMatrix:
 
    def __init__(self, size, queue_size):
       if size < 1:
@@ -24,25 +24,28 @@ class DirectionVector:
       for q in range(self.queue_size):
          self.frames_queue.put(dummy_frame)
 
-   def _compute_heat_matrix(self, new_frame):
+   def _update_heat_matrix(self, new_frame):
 
       #Compute heat matrix by by removing data from the last frame (divide)
       #and add new frame (multiply)
-      heat_matrix = (self.heat_matrix / self._get_oldest_frame()) * new_frame
+      self.heat_matrix = (self.heat_matrix / self._get_oldest_frame()) * new_frame
 
       #Add newest frame into our queue for the future
       self._add_new_frame(new_frame)
-   
-      return heat_matrix
 
    def compute_vector(self, new_frame):
+      #Get size
+      size = self.shape[0]
       
-      #Compute Newest heat matrix
-      heat_matrix = self._compute_heat_matrix(new_frame)
-      
+      #Update the heatmatrix with the new frame
+      self._update_heat_matrix(new_frame)
+
       #Get co-ordinate of highest value in heat_matrix
-      highest_cord = np.unravel_index(heat_matrix.argmax(), heat_matrix.shape)
+      highest_cord = np.unravel_index(self.heat_matrix.argmax(), self.heat_matrix.shape)
       
       #calculate the vector as (x1-x2),(y1-y2)
-      vector = [self.shape/2,self.shape/2] - highest_cord
+      #x1 and y1 being equivalent to origin
+      vector = ((size/2 - highest_cord[0])
+               ,(size/2 - highest_cord[1]))
+
       return vector
